refactor: rename `palette.h` to `p8.h`
Now that we have directories, this makes a lot more sense!
