+++ src/scene.c
@@ -1,5 +1,5 @@
 #include "context.h"
-#include "palette/palette.h"
+#include "palette/p8.h"
 #include "scene.h"
 #include "systems.h"
 #include <assert.h>
